add typescript

+++ controllers/ProfileController.ts
@@ -1,21 +1,23 @@
-import User from '../models/User.js'
-import Profile from '../models/Profile.js'
+import { User } from '../models/User'
+import Profile from '../models/Profile'
 import createError from 'http-errors'
 import sharp from 'sharp'
 import path from 'path'
 import fs from 'fs'
 import { validationResult } from 'express-validator'
+import { Response } from 'express'
+import { IRequest } from '../types/index'
 
 export class ProfileController {
   //GET PROFILE
-  async getProfile(req, res, next) {
+  async getProfile(req: IRequest, res: Response, next: any) {
     try {
       const userId = req.params.userId
-      const { photos } = await User.findById(userId)
+      const response = await User.findById(userId)
       const profile = await Profile.findOneAndUpdate(
         { userId },
         {
-          $set: { photos: { small: photos.small, large: photos.large } }
+          $set: { photos: { small: response?.photos.small, large: response?.photos.large } }
         },
         { new: true }
       )
@@ -26,11 +28,11 @@ export class ProfileController {
     }
   }
   //UPDATE PROFILE
-  async updateProfile(req, res, next) {
+  async updateProfile(req: IRequest, res: Response, next: any) {
     try {
       const err = validationResult(req)
-      if (!err.isEmpty()) { return next(createError(500, `${err.errors[0].msg}`)) }
-      const userId = req.user.id
+      if (!err.isEmpty()) { return next(createError(500, `${err.array()[0].msg}`)) }
+      const userId = req.user?.id
       if (!userId) { return next(createError(500, 'You are not autorizated')) }
       const profile = await Profile.findOneAndUpdate({ userId }, {
         $set: req.body
@@ -42,22 +44,22 @@ export class ProfileController {
     }
   }
   //GET STATUS
-  async getStatus(req, res, next) {
+  async getStatus(req: IRequest, res: Response, next: any) {
     try {
       const userId = req.params.userId
-      const { status } = await User.findById(userId)
+      const response = await User.findById(userId)
       // if (!profile) { return next(createError(500, 'User Id incorrect')) }
-      res.json(status);
+      res.json(response?.status);
     } catch (e) {
       return next(createError(500, 'Get status error'))
     }
   }
   //UPDATE STATUS
-  async updateStatus(req, res, next) {
+  async updateStatus(req: IRequest, res: Response, next: any) {
     try {
       const err = validationResult(req)
-      if (!err.isEmpty()) { return next(createError(500, `${err.errors[0].msg}`)) }
-      const id = req.user.id
+      if (!err.isEmpty()) { return next(createError(500, `${err.array()[0].msg}`)) }
+      const id = req.user?.id
       if (!id) { return next(createError(500, 'You are not autorizated')) }
       await User.findByIdAndUpdate(id, { status: req.body.status })
       // if (!profile) { return next(createError(500, 'User Id incorrect')) }
@@ -68,10 +70,10 @@ export class ProfileController {
     }
   }
   //FOLLOW USER
-  async followUser(req, res, next) {
+  async followUser(req: IRequest, res: Response, next: any) {
     try {
       const userId = req.params.userId
-      const currentUser = req.user.id
+      const currentUser = req.user?.id
       // if (!currentUser) { return next(createError(500, 'You are not autorizated')) }
       if (currentUser === userId) { return next(createError(500, 'You can not follow yourself')) }
       await User.findById(userId)
@@ -84,10 +86,10 @@ export class ProfileController {
     }
   }
   //UNFOLLOW USER
-  async unfollowUser(req, res, next) {
+  async unfollowUser(req: IRequest, res: Response, next: any) {
     try {
       const userId = req.params.userId
-      const currentUser = req.user.id
+      const currentUser = req.user?.id
       // if (!currentUser) { return next(createError(500, 'You are not autorizated')) }
       if (currentUser === userId) { return next(createError(500, 'You can not unfollow yourself')) }
       await User.findById(userId)
@@ -100,15 +102,15 @@ export class ProfileController {
     }
   }
   //IS CURRENT USER FOLLOWED
-  async isFollowed(req, res, next) {
+  async isFollowed(req: IRequest, res: Response, next: any) {
     try {
       const userId = req.params.userId
-      const currentUser = req.user.id
+      const currentUser = req.user?.id
       // if (!currentUser) { return next(createError(500, 'You are not autorizated')) }
       if (currentUser === userId) { return next(createError(500, 'You can not follow yourself')) }
       await User.findById(userId)
-      const { followedIds } = await User.findById(currentUser)
-      if (followedIds.includes(userId)) { return res.json(true) }
+      const response = await User.findById(currentUser)
+      if (response?.followedIds?.includes(userId)) { return res.json(true) }
       res.json(false)
     } catch (e) {
       // console.log(e);
@@ -116,21 +118,21 @@ export class ProfileController {
     }
   }
   //UPLOAD USER PHOTO
-  async putUserPhoto(req, res, next) {
+  async putUserPhoto(req: IRequest, res: Response, next: any) {
     try {
-      const currentUser = req.user.id
+      const currentUser = req.user?.id
       const fileName = 'img-' + currentUser + '.jpg'
       const filePath = path.resolve('uploads', fileName)
       const fileLink = process.env.REACT_APP_SERVER_API + fileName
-      sharp(req.file.path)
+      sharp(req.file?.path)
         .rotate()
         .resize(300, 300)
         .toFile(filePath, function (err, sharp) {
           if (err) {
-            fs.unlinkSync(req.file.path)
+            req.file && fs.unlinkSync(req.file.path)
             return next(createError(500, 'File format error'));
           }
-          fs.unlinkSync(req.file.path)
+          req.file && fs.unlinkSync(req.file.path)
           const data = {
             small: fileLink,
             large: fileLink
